Misc stuff and more bugfixes
This fixes a crash on the client, which I'm not sure when I caused, but it was caused by loading the server config on the client, so I need to find a workaround for that, so the server config isn't loaded on the client. Possibly by moving that injection using the server config to a dedicated server mixin.